Correct the value of C2 to 220nF in Eagle and BOM files
In the datasheet of LM4875 [1], Texas Instruments recommends
a 0.22uF (which is 220nF [2]) AC-coupling capacitor at the
audio input port [3]. However, in the Eagle and BOM files of
Gaboze Pocaio Round 2, the capacitor value was specified
as 220pF by accident. Experiments showed that using a 220pF
capacitor instead of 220nF causes an audible distortion
of the headphone output, which is characterized by the lack
of bass due to the RC high-pass filter action.

This commit corrects the component value C2 from 220pF to
220nF in Eagle schematic and board design files. Also, in
the BOM,

    C2,CAP CER 220PF 50V X7R 0805,C0805C221M5RACTU

has been replaced by

    C2,CAP CER 0.22UF 50V X7R 0805,C0805C224K5RACTU

Applying this fix to the circuit board fixes the audible
distortion and allows the audio quality to be restored
to TI's reference design.

It should be note that the Gerber files are not corrected,
since the committer doesn't have the project settings required
to create an identical set of Gerber files. The Gerber files
should be updated in a later commit by the original author.

[1] https://www.ti.com/lit/ds/symlink/lm4875.pdf

[2] https://en.wikipedia.org/wiki/Metric_prefix

[3] See Figure 1.

Signed-off-by: Tom Li <tomli@tomli.me>